Product Mange And Component Add

+++ database/migrations/2020_05_14_141427_create_product_bags_table.php
@@ -19,13 +19,13 @@ class CreateProductBagsTable extends Migration
             $table->string('pro_brand')->nullable();
             $table->integer('pro_type')->nullable();
             $table->string('pro_bar_code')->nullable();
-            $table->integer('pro_retail_price_per_kg')->nullable();
-            $table->integer('pro_retail_price_bag')->nullable();
-            $table->integer('pro_whole_sale_price_per_kg')->nullable();
-            $table->integer('pro_whole_sale_price_bag')->nullable();
-            $table->integer('pro_purchase_price_per_kg')->nullable();
-            $table->integer('pro_purchase_price_bag')->nullable();
-            $table->integer('pro_total_weight_in_bag')->nullable();
+            $table->double('pro_retail_price_per_kg')->nullable();
+            $table->double('pro_retail_price_bag')->nullable();
+            $table->double('pro_whole_sale_price_per_kg')->nullable();
+            $table->double('pro_whole_sale_price_bag')->nullable();
+            $table->double('pro_purchase_price_per_kg')->nullable();
+            $table->double('pro_purchase_price_bag')->nullable();
+            $table->double('pro_total_weight_in_bag')->nullable();
             $table->integer('pro_shop_id')->nullable();
             $table->integer('add_or_changeBy')->nullable();
